Resilience4j and Hystrix separate auto configuration classes
